build: Add continue as a submodule
This extension is a downstream extension from the
continuedev/continue project.

In order to leverage that project without hard forking
and diverging, this commit adds it as a submodule.

In this way we should be able to stay close to upstream,
while still being able to formfit the project around
Granite models.