build: Add stubs for VS Code tasks
In order to integrate with VS Code, the repository needs to provide
tasks to execute and build and run time. These tasks are defined
by the files tasks.json and launch.json respectively.

For now, this commit creates empty stubs for these files, which will
get populated in a future commit.