build: Add "Install Dependencies" task
Before we can build the project we need a way to install
its dependencies.

This commit adds:

- A new "Install Dependencies" task for VSCode to use, which calls
- An "install-dependencies" script definition in package.json to call
- An install-dependencies.js script to run
- "npm install" to install the dependencies specified in package.json
  and copy data files (like tree sitter wasms) to the install root.
- Also use npm to install arch specific dependencies into a staging
  area which will be useful later for building VSIX packages for other
  platforms.

Contributions by Owen Taylor to:
- Ensure the command output streams in real time using his new
  runCommand utility function.
- Tweak the output paths somewhat to closer match the way upstream
  continue does it.

build: Provide way to stage arch specific node_modules

In order to generate a VSIX package, we need to have node
modules for platforms that may differ from the build host.

This commit augments the install-dependencies script to allow
it to pull down node modules for other platforms to a staging
area, where they can later be plucked up to use during VSIX
packaging.