build: Add script for dynamically forking continue codebase
We want to work closely with continuedev upstream, but we also
need to make changes that are counter to their goals:

 - Make opinionated choices to streamline on-ramping
 - In particular, Granite.Code is supposed to showcase the Granite
   family of models, so we need model configuration to put those
   models front and center
 - Granite.Code is also supposed to be deliberately local only, so
   we need to make changes to ensure the user is confident of that.

As a middle ground between completely forking and not getting changes
in that are essential to the projects identity, we're going to
experiment with a "light fork" that keeps our changes on top, but
keeps the upstream continue source code close by.
by.

This commit partly achieves that by copying bits of the continue
codebase we want to leverage from upstream to the toplevel source
dir. At the time of the copy it does certain post-processing to
rename things, so Granite.Code gets its own identity in the process.

Contributions by Owen Taylor to transform the extension id and to
handle multiple substituions for the same string.