build: Add a script to update package.json from Continue's package.json
package.json is where certain actions used by the VSCode extension are
defined. We want these actions to match what continuedev upstream does
for the most part, so we would ideally generate them at build time.

But, package.json is also used by npm for starting the build, so it
must exist ahead of time in version control, so we can't really
generate it at build time.

This commit adds a script update-package-json.js to sync package.json
with the latest upstream.

It won't get run as part of the build pipeline, but just run
periodically as upstream changes.