build: Add a runCommand() utility function
When scripts are run in VSCode the output is comingled into
one pane.

For clarity, it's useful to know if the output is coming directly
from our scripts or is, e.g., coming from npm.

This commit adds a utility function, runCommand, to run commands
with each line annotated with a designated prefix.