build: Provide way to stage arch specific node_modules
In order to generate a VSIX package, we need to have node
modules for platforms that may differ from the build host.

This commit augments the install-dependencies script to allow
it to pull down node modules for other platforms to a staging
area, where they can later be plucked up to use during VSIX
packaging.