build: Generate pre-release packages by default
We aren't ready for a real release yet, but we do need
vsce output generated with the pre-release flag so it can
be published early on the marketplace.

This commit makes the packages get generated with the
pre-release flag by default, and changes --pre-release
to --releast to drop the flag.