build: Copy sqlite3 modules for all platforms
The configuration entry for installing sqlite3 has this condition:

platforms: [`${process.platform}-${process.arch}`]

Meaning, it only gets run for the current platform, which breaks
packages created with --target to a different platform.

This commit changes it to:

platforms: ["linux-x64", "linux-arm64", "darwin-arm64", "win32-x64"],

So it runs for all platforms.