build: Make platforms list clearer in create-package
At the moment the list of platforms is implicitly defined by
the list of platforms that have variables.

It's important to have a single source of truth for the list,
and using the variables array for that source of truth is not
obvious when casually reading the code.

This commit changes the variables array to instead be a platforms
object, keyed off the platform name. Doing it this way, makes the
platform definitions very explicit, which is more fitting for a
single source of truth.