install-dependencies: Structure more like create-package
create-package.js and install-dependencies have similar
logic for handling platform expansion, but are slightly
divergent.

At the moment the install-dependencies logic is failing to
install sqlite3 correctly on all platforms because of
a difference of opinion between configuration and code on
what should happen if no platforms are specified in
configuration.

This commit makes install-dependencies use the same
logic as create-package, where the platforms list
is defined explicitly, platform aliases are separated out,
missing platforms key in a configuration entry means
"all platforms" if there are arch specific substitutions,
etc, etc.