build: Add script for bumping and tagging next prerelease
Every night that there is change, we want to push a new prerelease
to the marketplace.

In order to do this we need to update the version of the package.

We don't actually want to commit that update, though, because it's
not a real release. We should tag it in git, to mark when the
last package was built.

This commit adds a script that uses git tags as the source of truth
for which version to use. It adopts an even/odd minor number
release/prerelease versioning scheme.