ci: Fetch tags when cloning repo
The publishing logic depends on tags being in the cloned
repository to know which pre-release version to assign to
the output package.

Unfortunately, GitHub by default does a shallow clone with
no branches or tags.

This commit adds `fetch-tags: true` to ensure the tags are
fetched at clone time.

Suggestion by Owen Taylor.