scripts/tag-prerelease: Rename to tag-release.js
This script will ultimately get updated to handle doing non-prelease
releases as well.

As a step in that direction, this commit renames the script to drop the
"pre" from its name.

Also, this commit updates package.json to reference the new name, and to adds a
(currently ignored) --snapshot parameter, which will later be used to
disambiguate the two modes of the script.