scripts/tag-release: Support non-snapshot releases
We're getting close to the point where we want to do a regular
stable release, but our machinery currently only supports generating
nightlies.

This commit updates the tag-release script to be able to tag proper
releases, and updates the github workflow to allow a way to kick
the process off using a new on-demand.yaml workflow.

This commit requires writing out package.json from the tag-release
script for non-snapshot releases, so for consistency, it moves the
package.json writing for pre-releases from publish.yaml to
tag-release.js as well.