workflows: Cleave publishing logic to reusable file
We're going to need to support on-demand publishing in
order to support doing proper releases.

There's a lot of logic in publish.yaml that can be reused
between nightly and on-demand workflows.

As an initial step toward leveraging that reusable logic,
this commit separates the cron aspect off
to a new nightly.yaml, and relegates publish.yaml to be
a support file that gets included by nightly.yaml.