scripts/tag-release: Read package.json from right place
The code assumes tag-release script is run from within the
scripts/ directory, which is not actually how we do it in practice.

In practice we do

node scripts/tag-release.js

which erroneously makes the code look for package.json outside of
the source repository.

This commit fixes the issue by looking for package.json in the parent
directory of the script itself, so it doesn't matter where the user
runs the script from.