workflows: Ensure release packages aren't built with --pre-release
For a package to be considered a release package, it needs to be
built without --pre-release and published without --pre-release.

At the moment, we're publishing okay, but neglecting to take
--pre-release off the package building part.

This commit fixes that, by making create-package take --snapshot
instead of --release as an argument, and then changing the workflow
to specifically add --snapshot for snapshots.